requirements.txt
pymupdf
pymupdf4llm
google-genai
fastapi
uvicorn
python-multipart
python-dotenv
pillow
